moved conversion stuff to final place and created factory
