+++ transformer_lens/utilities/architectures.py
@@ -1,0 +1,20 @@
+SUPPORTED_ARCHITECTURES = [
+    "BertForMaskedLM",
+    "BloomForCausalLM",
+    "Gemma2ForCausalLM",
+    "GemmaForCausalLM",
+    "GPT2LMHeadCustomModel",
+    "GPT2LMHeadModel",
+    "OPTForCausalLM",
+    "GPTJForCausalLM",
+    "GPTNeoForCausalLM",
+    "GPTNeoXForCausalLM",
+    "LlamaForCausalLM",
+    "MistralForCausalLM",
+    "MixtralForCausalLM",
+    "Phi3ForCausalLM",
+    "PhiForCausalLM",
+    "Qwen2ForCausalLM",
+    "QWenLMHeadModel",
+    "T5ForConditionalGeneration",
+]
