+++ transformer_lens/weight_conversion/conversion_utils/conversion_steps/base_weight_conversion.py
